Fixed issue where mobile menu doesn't appear on short scrolls
Occurs mainly if someone enters a page where the anchor is less than 40 pixels away from the top of the page. Since we're smoothing out scrolling over a 40 pixel range, if this occurs, it doesn't detect the user hitting the top of the page where it should show the menu. This fix will ensure that if the top of the page is hit, the mobile menu always shows.